+++ .github/workflows/ci.yml
@@ -10,16 +10,18 @@ env:
   CARGO_TERM_COLOR: always
 
 jobs:
-  build:
-
+  stable:
     runs-on: ubuntu-latest
-
     steps:
-    - uses: actions/checkout@v2
-    - uses: actions-rs/toolchain@v1
-      with:
-        toolchain: 1.31
-    - run: cargo build --verbose
+    - uses: actions/checkout@v4
+    - uses: dtolnay/rust-toolchain@stable
     - run: cargo build --verbose --all-features
-    - run: cargo test --verbose
     - run: cargo test --verbose --all-features
+
+  msrv:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - uses: dtolnay/rust-toolchain@1.48.0
+    - run: cargo build --verbose
+    - run: cargo test --verbose
