+++ .gitignore
@@ -1,6 +1,5 @@
 /target/
 **/*.rs.bk
-Cargo.lock
 .vscode
 .idea
 .DS_Store
